pyproject.toml
[build-system]
build-backend = "uv_build"
requires = [ "uv-build" ]

[project]
name = "meal-planner"
version = "0.1.0"
description = "Recipe collector and weekly meal planner with automatic ingredient extraction"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14.2"
classifiers = [
  "Private :: Do Not Upload",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "email-validator>=2.3",
  "fastapi>=0.115",
  "firebase-admin>=7.1",
  "google-cloud-firestore>=2.19",
  "google-cloud-storage>=3.8",
  "google-genai>=1.60",
  "httpx>=0.27",
  "pillow>=10",
  "pydantic>=2.10",
  "python-dotenv>=1",
  "python-multipart>=0.0.22",
  "recipe-scrapers>=15",
  "uvicorn>=0.34",
  "watchdog>=6",
]

optional-dependencies.dev = [
  "deptry>=0.24",
  "meal-planner[test]",
  "pre-commit>=3.6",
  "pyproject-fmt>=2.11",
  "ruff>=0.11",
  "ty>=0.0.5",
]
optional-dependencies.test = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "pytest-cov>=4.1",
  "pytest-mock>=3.12",
  "pytest-xdist>=3.5",
  "respx>=0.22",
]

[dependency-groups]
dev = [
  "meal-planner[dev]",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py313" # Ruff doesn't support py314 yet, py313 is compatible

line-length = 120
indent-width = 4
cache-dir = ".cache/.ruff_cache"

format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "lf"
format.skip-magic-trailing-comma = true

lint.select = [
  "ALL", # include all the rules, including new ones
]
lint.ignore = [
  "ANN401",  # Any type (acceptable for dynamic types)
  "BLE001",  # Blind exception (acceptable with proper error handling)
  "COM812",  # Trailing comma conflicts with formatter
  "D",       # Docstring rules (enable if needed)
  "E501",    # Line too long (handled by formatter)
  "INP001",  # Implicit namespace package
  "PLC0415", # Import not at top-level (needed for optional dependencies)
  "S311",    # Random for non-crypto (not used for security purposes)
  "T201",    # Print statements (enable if prints should be avoided)
  "TRY300",  # Consider else block (style preference)
]

lint.per-file-ignores."api/routers/**/*.py" = [
  "FBT001", # Boolean positional argument (FastAPI Query parameters)
  "FBT002", # Boolean default positional argument (FastAPI Query parameters)
]

lint.per-file-ignores."api/services/recipe_enhancer.py" = [
  "PLR0913", # enhance_recipe has many params from household config passthrough
]

lint.per-file-ignores."scripts/**/*.py" = [
  "ANN201",  # Missing return type annotation ok in CLI scripts
  "C901",    # Complex functions ok in CLI scripts (main, process_batch)
  "E402",    # Module-level import not at top (needed for sys.path manipulation)
  "F841",    # Unused variable ok (sometimes used for debugging)
  "FBT003",  # Boolean positional value ok (Firestore FieldFilter)
  "PERF401", # List comprehension optimization not needed for Firestore iteration
  "PLR0911", # Too many return statements ok in CLI argument parsing
  "PLR0912", # Too many branches ok in CLI argument parsing
  "PLR0915", # Too many statements ok in batch processing scripts
  "PLR2004", # Magic numbers ok in CLI arg parsing
]

lint.per-file-ignores."tests/**/*.py" = [
  "ANN001",  # Missing type annotation for function argument
  "ARG001",  # Unused function argument (fixtures used for side effects)
  "ARG002",  # Unused method argument (fixtures used for side effects)
  "PLR2004", # Allow magic values in tests
  "S101",    # Allow assert in tests (required for pytest)
]

lint.per-file-ignores."tools/**/*.py" = [
  "ANN201",  # Missing return type annotation ok in CLI tools
  "C901",    # Complex functions ok in CLI tools (interactive menus)
  "PERF401", # List comprehension optimization not needed for Firestore iteration
  "PLR0912", # Too many branches ok in CLI argument parsing
  "PLR0915", # Too many statements ok in CLI tools
  "PLR2004", # Magic numbers ok in CLI tools
]

lint.flake8-quotes.inline-quotes = "double"
lint.isort.combine-as-imports = true
lint.isort.split-on-trailing-comma = false

[tool.deptry]
known_first_party = [ "meal-planner" ]
pep621_dev_dependency_groups = [ "dev" ]
ignore = [ "DEP002" ]
extend_exclude = [ ".venv/", "functions/", "scripts/", "tools/" ]

[tool.pytest.ini_options]
testpaths = [ "tests" ]
python_files = [ "test_*.py" ]
python_functions = [ "test_*" ]
filterwarnings = [
  "error",
  # MagicMock in tests can trigger this when Pydantic validates async clients
  "ignore::pydantic.warnings.ArbitraryTypeWarning",
]
addopts = "-W error --strict-markers"
markers = [
  "asyncio: mark a test as an async test (auto-discovered by pytest-asyncio)",
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = [ "api" ]
branch = true
parallel = true
omit = [
  "tests/*",
  "*/__pycache__/*",
]

[tool.coverage.report]
fail_under = 95
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
]
